requirements.txt
flask
flask_sqlalchemy
pillow
torch
torchvision
numpy
scikit-learn
gunicorn
